extract EventInfo component
