+++ client/src/components/EventPage/index.tsx
@@ -1,26 +1,13 @@
 import React, { useState, useEffect } from 'react';
-import Grid from '@material-ui/core/Grid';
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableHead from '@material-ui/core/TableHead';
-import TableRow from '@material-ui/core/TableRow';
-import TableContainer from '@material-ui/core/TableContainer';
-import Paper from '@material-ui/core/Paper';
 import AppBar from '@material-ui/core/AppBar';
 import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
-import { makeStyles } from '@material-ui/core/styles';
-
 import Event from '../../model/event';
-import EventClass from '../../model/event-class';
 import { match } from "react-router-dom";
-import moment from 'moment';
-
-import RegistrationInfo from './RegistrationInfo';
 import ParticipantList from '../ParticipantList';
 import Firebase from '../Firebase';
-import HeaderBar from "../headerbar/HeaderBar";
+import HeaderBar from '../headerbar/HeaderBar';
+import EventInfo from './EventInfo';
 
 interface MatchParams {
     eventId: string
@@ -49,73 +36,12 @@ function TabPanel(props: TabPanelProps) {
 const EventPage: React.FC<Props> = (props: Props) => {
 
     const [event, setEvent] = useState<Event>(new Event("", "", "", "", new Date(), new Date(), new Date(), false, false, [], []));
-
     const [tabIndex, setTabIndex] = useState(0);
 
     useEffect(() => {
         const eventId = props.match.params.eventId;
         return Firebase.subscribeEvent(eventId, setEvent);
     }, [props.match]);
-
-    const useStyles = makeStyles({
-        infoTable: {
-            '& td': {
-                verticalAlign: "top"
-            },
-            '& ul': {
-                margin: "0",
-                padding: "0 0 0 16px"
-            }
-        },
-        noMargin:  {
-            marginTop: "0",
-        }
-      });
-
-    const classes = useStyles({});
-    const description = {__html: event.description}
-    const infoTab = (
-        <React.Fragment>
-            <Grid container spacing={2}>
-                <Grid item xs={12} md={6}>
-                    <TableContainer component={Paper}>
-                        <Table className={classes.infoTable}>
-                            <TableBody>
-                                <TableRow><TableCell>Arrangement:</TableCell><TableCell>{event.name}</TableCell></TableRow>
-                                <TableRow><TableCell>Dato:</TableCell><TableCell>{moment(event.startTime).format("DD. MMM YYYY")}</TableCell></TableRow>
-                                <TableRow><TableCell>Øvelse:</TableCell><TableCell>{event.eventType}</TableCell></TableRow>
-                                <TableRow><TableCell>Første start:</TableCell><TableCell>{moment(event.startTime).format("HH:mm")}</TableCell></TableRow>
-                                <TableRow><TableCell>Arrangementsinfo:</TableCell><TableCell><span className={classes.noMargin} dangerouslySetInnerHTML={description}/></TableCell></TableRow>
-                                <RegistrationInfo event={event} />
-                            </TableBody>
-                        </Table>
-                    </TableContainer>
-                </Grid>
-                <Grid item xs={12} md={6}>
-                    <TableContainer component={Paper}>
-                        <Table size="small">
-                            <TableHead>
-                                <TableRow>
-                                    <TableCell>Klasse</TableCell>
-                                    <TableCell>Løype</TableCell>
-                                    <TableCell>Beskrivelse</TableCell>
-                                </TableRow>
-                            </TableHead>
-                            <TableBody>
-                                {event.eventClasses.map((ec:EventClass, idx:number) => (
-                                    <TableRow key={ec.name}>
-                                        <TableCell>{ec.name}</TableCell>
-                                        <TableCell>{ec.course}</TableCell>
-                                        <TableCell>{ec.description}</TableCell>
-                                    </TableRow>
-                                ))}
-                            </TableBody>
-                        </Table>
-                    </TableContainer>
-                </Grid>
-            </Grid>
-        </React.Fragment>
-    );
 
     const handleTabChange = (event: React.ChangeEvent<{}>, newValue: number) => {
         setTabIndex(newValue);
@@ -131,14 +57,13 @@ const EventPage: React.FC<Props> = (props: Props) => {
                 </Tabs>
             </AppBar>
             <TabPanel value={tabIndex} index={0}>
-                {infoTab}
+                <EventInfo event={event}/>
             </TabPanel>
             <TabPanel value={tabIndex} index={1}>
                 <ParticipantList event={event}/>
             </TabPanel>
         </React.Fragment>
     );
-    
 }
 
 export default EventPage;
